agent_apps/aws_strands_agent/requirements.txt
strands-agents[openai]
strands-agents
strands-agents-tools
streamlit
